Rearrange the code for nicer support
