+++ components/job-orchestration/job_orchestration/executor/compress/celeryconfig.py
@@ -5,16 +5,12 @@ from job_orchestration.scheduler.constants import QueueName, TASK_QUEUE_HIGHEST_
 # Worker settings
 # Force workers to consume only one task at a time
 worker_prefetch_multiplier = 1
-imports = [
-    "job_orchestration.executor.compress.fs_compression_task",
-    "job_orchestration.executor.compress.fs_to_s3_compression_task"
-]
+imports = ["job_orchestration.executor.compress.fs_compression_task"]
 
 # Queue settings
 task_queue_max_priority = TASK_QUEUE_HIGHEST_PRIORITY
 task_routes = {
-    "job_orchestration.executor.compress.fs_compression_task.compress": QueueName.COMPRESSION,
-    "job_orchestration.executor.compress.fs_to_s3_compression_task.compress": QueueName.COMPRESSION,
+    "job_orchestration.executor.compress.fs_compression_task.compress": QueueName.COMPRESSION
 }
 task_create_missing_queues = True
