--- components/job-orchestration/job_orchestration/executor/compress/utils.py
@@ -1,61 +1,0 @@
-import pathlib
-import subprocess
-
-from clp_py_utils.clp_config import StorageEngine
-from job_orchestration.scheduler.job_config import ClpIoConfig, PathsToCompress
-import os
-import yaml
-
-def make_clp_command(
-    clp_home: pathlib.Path,
-    archive_output_dir: pathlib.Path,
-    clp_config: ClpIoConfig,
-    db_config_file_path: pathlib.Path,
-):
-    path_prefix_to_remove = clp_config.input.path_prefix_to_remove
-
-    # fmt: off
-    compression_cmd = [
-        str(clp_home / "bin" / "clp"),
-        "c", str(archive_output_dir),
-        "--print-archive-stats-progress",
-        "--target-dictionaries-size", str(clp_config.output.target_dictionaries_size),
-        "--target-segment-size", str(clp_config.output.target_segment_size),
-        "--target-encoded-file-size", str(clp_config.output.target_encoded_file_size),
-        "--db-config-file", str(db_config_file_path),
-    ]
-    # fmt: on
-    if path_prefix_to_remove:
-        compression_cmd.append("--remove-path-prefix")
-        compression_cmd.append(path_prefix_to_remove)
-
-    # Use schema file if it exists
-    schema_path: pathlib.Path = clp_home / "etc" / "clp-schema.txt"
-    if schema_path.exists():
-        compression_cmd.append("--schema-path")
-        compression_cmd.append(str(schema_path))
-
-    return compression_cmd
-
-
-def make_clp_s_command(
-    clp_home: pathlib.Path,
-    archive_output_dir: pathlib.Path,
-    clp_config: ClpIoConfig,
-    db_config_file_path: pathlib.Path,
-):
-    # fmt: off
-    compression_cmd = [
-        str(clp_home / "bin" / "clp-s"),
-        "c", str(archive_output_dir),
-        "--print-archive-stats",
-        "--target-encoded-size", str(clp_config.output.target_segment_size + clp_config.output.target_dictionaries_size),
-        "--db-config-file", str(db_config_file_path),
-    ]
-    # fmt: on
-
-    if clp_config.input.timestamp_key is not None:
-        compression_cmd.append("--timestamp-key")
-        compression_cmd.append(clp_config.input.timestamp_key)
-
-    return compression_cmd
