+++ components/job-orchestration/job_orchestration/executor/search/fs_search_task.py
@@ -1,27 +1,19 @@
-import errno
 import os
-import threading
-from shutil import rmtree
 import signal
 import subprocess
 import sys
-import time
 from pathlib import Path
-from threading import Thread
 from typing import Any, Dict
 
 from celery.app.task import Task
 from celery.utils.log import get_task_logger
 from clp_py_utils.clp_config import StorageEngine
 from clp_py_utils.clp_logging import set_logging_level
+
 from job_orchestration.executor.search.celery import app
+from job_orchestration.executor.utils import post_cleanup, setup_reader_archive_directory
 from job_orchestration.scheduler.job_config import SearchConfig
 from job_orchestration.scheduler.scheduler_data import SearchTaskResult
-from remote_fuse_layer.s3 import (
-    S3MountConfig,
-    clp_s3_fuse_sequential_read_thread_method
-)
-from job_orchestration.executor.credentials import HARDCODED_S3_CONFIG
 
 # Setup logging
 logger = get_task_logger(__name__)
@@ -90,141 +82,6 @@ def make_clp_s_command(
     return search_cmd
 
 
-def _get_mount_cache_path(mount_path: Path) -> Path:
-    return mount_path.parent / ("." + mount_path.name)
-
-
-def _remove_and_mkdir(dir_path: Path, is_fuse_mount: bool = False) -> None:
-    try:
-        if dir_path.exists():
-            rmtree(dir_path)
-    except OSError as e:
-        if is_fuse_mount and e.errno == errno.ENOTCONN:
-            # ENOTCONN corresponds to
-            # "transport endpoint is not connected: <mount_point>"
-            subprocess.run(["fusermount", "-u", str(dir_path)])
-            if dir_path.exists():
-                rmtree(dir_path)
-        else:
-            raise e
-
-    dir_path.mkdir(parents=True)
-
-
-def _get_s3_mount_config(s3_yaml_config: Dict[str, Any]) -> S3MountConfig:
-    return S3MountConfig(
-        access_key_id=s3_yaml_config["access_key_id"],
-        secret_access_key=s3_yaml_config["secret_access_key"],
-        s3_path_prefix_str=s3_yaml_config["s3_path_prefix"],
-        endpoint_url=s3_yaml_config["endpoint_url"],
-        s3_path_prefix_to_remove_from_mount=None,
-    )
-
-
-def _await_fuse_mount(mount_path: Path) -> None:
-    logger.info(f"Waiting for {mount_path} to mount...")
-    while not os.path.ismount(mount_path):
-        time.sleep(0.01)
-    logger.info(f"{mount_path} mounted.")
-
-
-def _unmount_fuse(mount_path: Path, fuse_thread: Thread) -> None:
-    logger.info(f"Unmounting {mount_path}...")
-    subprocess.run(
-        ["fusermount", "-u", str(mount_path)],
-        close_fds=True,
-        stdout=subprocess.DEVNULL,
-        stderr=subprocess.STDOUT,
-    )
-
-    logger.info("Waiting for the FUSE thread to exit...")
-    fuse_thread.join()
-    logger.info("FUSE thread exited.")
-
-    rmtree(mount_path)
-    rmtree(_get_mount_cache_path(mount_path))
-
-
-def _create_reader_fuse_thread(
-    s3_reader_config: Dict[str, Any], mount_path: Path
-) -> threading.Thread:
-    # Load reader S3 config
-    _remove_and_mkdir(mount_path, is_fuse_mount=True)
-
-    # Create a real on disk cache
-    reader_mount_cache_path = _get_mount_cache_path(mount_path)
-    _remove_and_mkdir(reader_mount_cache_path)
-
-    s3_reader_mount_config = _get_s3_mount_config(s3_reader_config)
-
-    # Create the thread as a daemon so that it exits immediately when the main
-    # thread is cancelled
-    reader_fuse_thread = threading.Thread(
-        target=clp_s3_fuse_sequential_read_thread_method,
-        args=(
-            s3_reader_mount_config,
-            mount_path,
-            reader_mount_cache_path,
-            s3_reader_config["max_file_size"],
-        ),
-        daemon=True,
-    )
-
-    return reader_fuse_thread
-
-
-def setup_archive_directory(
-    job_id: str,
-    task_id: str
-) -> Dict[str, Any]:
-
-    # # for the normal file system flow
-    # archive_output_dir = os.getenv("CLP_ARCHIVE_OUTPUT_DIR")
-    # if archive_output_dir:
-    #     return {
-    #         "archive_output_path": Path(archive_output_dir)
-    #     }
-
-    # for the S3 Flow
-    s3_reader_config = HARDCODED_S3_CONFIG
-    # 4GB should be a very safe guess.
-    s3_reader_config["max_file_size"] = 4 * 1024 * 1024 * 1024
-
-    # A temp hack for access key id and secret
-    os.environ["AWS_ACCESS_KEY_ID"] = s3_reader_config["access_key_id"]
-    os.environ["AWS_SECRET_ACCESS_KEY"] = s3_reader_config["secret_access_key"]
-
-    # generate reader_mount path
-    reader_mount_path = Path("/") / f"reader_mount_{job_id}_{task_id}"
-
-    # Launch the reader mount thread
-    reader_fuse_thread = _create_reader_fuse_thread(s3_reader_config, reader_mount_path)
-    reader_fuse_thread.start()
-    logger.info("Creating FUSE mount...")
-    _await_fuse_mount(reader_mount_path)
-
-    # point the storage dir to mounted path
-    return {
-        "archive_output_path": reader_mount_path,
-        "fuse_thread": reader_fuse_thread
-    }
-
-
-def post_search_cleanup(
-    setup_config: Dict[str, Any]
-) -> None:
-    # # Normal flow, do nothing
-    # if os.getenv("CLP_ARCHIVE_OUTPUT_DIR"):
-    #     return
-
-    # For the S3 Flow
-    _unmount_fuse(
-        setup_config["archive_output_path"],
-        setup_config["fuse_thread"]
-    )
-
-
-
 @app.task(bind=True)
 def search(
     self: Task,
@@ -249,9 +106,8 @@ def search(
     logger.info(f"Started task for job {job_id}")
 
     search_config = SearchConfig.parse_obj(search_config_obj)
-    setup_result = setup_archive_directory(job_id, task_id)
-    archive_directory = setup_result["archive_output_path"]
-    archive_path = archive_directory / archive_id
+    archive_dir, setup_info = setup_reader_archive_directory(job_id, task_id)
+    archive_path = archive_dir / archive_id
 
     if StorageEngine.CLP == clp_storage_engine:
         search_cmd = make_clo_command(
@@ -264,7 +120,7 @@ def search(
     elif StorageEngine.CLP_S == clp_storage_engine:
         search_cmd = make_clp_s_command(
             clp_home=clp_home,
-            archives_dir=archive_directory,
+            archives_dir=archive_dir,
             archive_id=archive_id,
             search_config=search_config,
             results_cache_uri=results_cache_uri,
@@ -296,7 +152,7 @@ def search(
             os.waitpid(search_proc.pid, 0)
             logger.info(f"Cancelling search task.")
             # Post search setup
-            post_search_cleanup(setup_result)
+            post_cleanup(archive_dir, setup_info)
         # Add 128 to follow convention for exit codes from signals
         # https://tldp.org/LDP/abs/html/exitcodes.html#AEN23549
         sys.exit(_signo + 128)
@@ -316,7 +172,7 @@ def search(
         logger.info(f"Search task completed for job {job_id}")
 
     # post search
-    post_search_cleanup(setup_result)
+    post_cleanup(archive_dir, setup_info)
 
     # Close log files
     clo_log_file.close()
